Update INF4000.md
The codes are about a side question of INF4000, from the topic of INF6027, focuses on the distribution characteristics of different emotion categories (such as anger, sadness, happiness, etc.) in news headlines and their changing trends over time and categories. There are four different kinds of visualisations to illustrate the topic.